requirements.txt
gradio>=3.50.0
GitPython>=3.1.40
rich>=13.7.0
transformers>=4.41.0
torch>=2.2.0
hf_xet